Correct data; Add Penalty to model
1. Correct data
For PredictorWithHis, the data has been corrected by using all the
resumes that has an active day in the last 5 days. The volume of
training and test data is enlarged after this correction.

2. Add Penalty to model
Currently RFWithHis has extremely good precision but not so good recall
rate. And RFLR improves the result a little bit and more importantly,
  provides a chance to mediate the significance of precision and recall.